(GH-2006) Add Powershell Module Manifest Template
Adds the `PuppetBolt.psd1` manifest file as a erb template that is built along with the `PuppetBolt.psm1` module file. This completes managing the PowerShell module creation process.

This adds the functions to export in the `PuppetBolt.psd1` manifest file and removes the `Export-ModuleMember` call in the `PuppetBolt.psm1` file to align with PowerShell best practices.

It also sets the module version to the same version of bolt, so that if a user looks at the module version it reports the version of bolt installed. This is important because the functionality of this PowerShell module is related 1:1 to the Bolt version is ships with.

!feature

* **Add Powershell Module Manifest Template**
     ([#2006](#2006))

     Add Powershell Module Manifest Template to be built along with main module file.