(GH-1993) Remove -E from bash commands
Previously we passed `-E` to the sudo command built in the bash shell if
the operation required environment variables to be passed. Prior to #1899
this was needed because we set environment variables when invoking sudo
and required it to preserve them so they would be visible to the command
being invoked. We now pass the environment variables as arguments to
sudo, making the flag unnecessary and raising errors for users who don't
have permissions to preserve the environment.

Closes #1993

!feature

* **Environment preservation permission no longer required when using run-as** ([1993](https://github.com/puppetlabs/bolt/issues/1993))

  Bolt no longer passes the `-E` flag to sudo when building commands
  using 'run-as', which allows users who do not have permission to use the
  flag to use Bolt.