(GH-1980) Add --env-var cli flag
This adds the `--env-var` cli flag, which accepts a string that contains
an equals sign and parses it into a key-value pair to set as an
environment variable on the target. This flag is only available when
running commands or scripts, and can be specified multiple times with
different environment variables. If the same key is specified multiple
times the last instance will take precedence. This doesn't interact with
`_env_var` specified in plans since the flag can only be used when
running commands and scripts.

Closes #1980

!feature

* **Specify remote environment variables using --env-var** ([1980](https://github.com/puppetlabs/bolt/issue/1980))

  Users can now set environment variables on targets when running
  commands and scripts using the CLI flag `--env-var`.