(GH-1980) Add secure_env_vars plan
This adds a a new plan `secure_env_vars` to Bolt's internal modulepath.
The plan accepts either a command or script, and errors if both or
neither is given, as well as targets. It reads environment variables as
JSON from a special local environment variable `BOLT_ENV_VARS` on the
Bolt controller and passes those environment variables to the command or
script.

This allows users to specify sensitive data as environment variables on
the Bolt controller and refer to those variables from a command or
script on remote systems.

!feature

* **Add secure_env_vars plan** ([GH-1980](https://github.com/puppetlabs/bolt/issues/1980))

  The new builtin Bolt plan `secure_env_vars` reads JSON from special
  environment variable `BOLT_ENV_VARS` and passes that hash to either
  run_command or run_script.