(GH-1982) Handle WinRM connection loss
Previously, losing connection to WinRM (for instance, due to the
service being stopped or restarted) would cause a stacktrace and a
deadlock. This happened because we manage the output from the command in
a thread which wasn't handling exceptions. The stack trace came from the
exception bubbling up, with the default behavior of Thread being to
print all exceptions that escape the thread. The deadlock came from our
failure to close the out/err streams that the shell was waiting on.

We now ensure the out/err streams are always closed and we disabled the
"report_on_exception" behavior to avoid an unnecessary stack trace, as
our own code will handle the exception at the next level up.

!bug

  * **Gracefully handle WinRM connection loss**
     ([#1982](https://github.com/puppetlabs/bolt/issues/1982))

    Bolt will now detect WinRM connection loss and return an error
    rather than printing a stacktrace and deadlocking.