(bug) Save plan functions reference to plan_functions.md
This updates the `docs:function_reference` task to save the generated
markdown doc to `documentation/plan_functions.md` instead of
`documentation/reference.md`. This path was erroneously changed as part
of the updates to Bolt's rake tasks in #2011.

This also updates the `reference.md.erb` template to be
`plan_functions.md.erb`, in line with the other templates that match the
name of the markdown files they generate.

!no-release-note