(GH-1997) Add tests for JSON schemas
This adds tests to ensure that the config options hashes include
`:type` and `:description` keys, which are required to correctly
generate Bolt's JSON schemas.

It also includes a workflow for generating the JSON schemas and
checking if there were any changes from the schemas that were committed.
If there are changes, the workflow fails and notifies the user that they
should run the `schemas:all` Rake task and commit the changes.

!no-release-note