(GH-1841) Prefer bolt-*.yaml and inventory.yaml to bolt.yaml in documentation
This modifies the documentation to prefer bolt-project.yaml,
bolt-defaults.yaml, and inventory.yaml over bolt.yaml. This mostly
includes instructing users to create a `bolt-project.yaml` instead of a
`bolt.yaml` file to create a Bolt project (both of which work at this
time), and specifying which file certain configuration options should
belong it. It doesn't update the documented default files or remove
bolt.yaml documentation, as bolt.yaml is still technically default and
has not been removed.

Closes #1841

!no-release-note