(GH-2049) Remove parameter help replacement
The code to parse the parameter description fields was copied from the bolt documentation rake tasks which generate text that may be used on a website and cannot have some special characters in it. We do not need to worry about this inside the PowerShell cmdlet help blocks, so this replacement is unnecessary.

!bug

  * **Fix parameter description parsing** ([#2049](https://github.com/puppetlabs/bolt/issues/2049))

    Removes encoding HTML special characters from the parameter description fields