(GH-2049) Add global bolt parameter mapping
Modify the cmdlet template and backing rake task code to create a global bolt parameter to powershell parameter mapping, instead of one mapping per powershell cmdlet. Since all bolt parameters are unique and are not implemented differently in each subcommand or action, this works without risking redefining parameters.

This also reducs the line count of the generated PowerShell module by hundreds of lines, as it does not duplicate the same set of parameters over and over.

!no-release-note