(GH-2049) Capitalize parameters in CamelCase
Capitalize all generated parameters in CamelCase form. This is a PowerShell coding style best practice, and expected in a properly formatted PowerShell script. This increases readability and comprehension of the generated code, but more importantly makes the parameters look as a PowerShell user would expect them inside the PowerShell commandline and help system.

This does not affect the bolt parameter to powershell parameter global mapping as that is a case insensitive match.