(GH-1996) Add documentation on setting up VS Code for Bolt
This adds a new page to Bolt's documentation that describes setting up
VS Code for use with Bolt. It describes how to use the JSON schemas to
validate configuration and inventory files and information about using
the Puppet VS Code extension to develop Bolt content.

This also removes the `bolt_transport_definitions.json` file which was
previously used by the schemas before they were generated. Since
this file is no longer used by the schemas, it can be deleted.

!no-release-note