(GH-1981) Add dir::children plan function
This adds a new builtin Bolt module `dir` that contains utilities for
interacting with directories during a Bolt plan, analagous to Ruby's
`Dir` class. It currently has just one function, `children`, which lists
the contents of a given directory and returns an ordered list of bare
filenames. This is used in the `facts::external` plan to get the list of
scripts to run to generate facts from.

The function accepts either an absolute path, a path relative to the
root of the active Bolt project, or a module and subpath. All of these
are only accepted as strings. If none are found, an error is raised.

!feature

* **New dir::children plan function**

  New plan function `dir::children` returns an array containing all of
  the filenames in the given directory, similar to Ruby's `Dir.children()`