(maint) Catch changes to bolt_options_parser
This github action should have failed a previous PR because we check that the 'correct' number of parameters are generated for each command (as a safety measure) and we added envvars to bolt command. I think this happened because I did not put in a check for the `bolt_option_parser.rb` file.

!no-release-note