(bug) Catch existing file errors when downloading files
This catches existing file errors that are raised when creating the
destination directory for file downloads in the
`Bolt::Executor#download_file` method. Previously, if any portion of the
destination directory was an existing file, Bolt would error and emit a
backtrace. This catches those errors and adds additional information
that the destination directory could not be created.

Any similar errors raised when files are being downloaded by the
individual transports are already caught and handled by the transport.

!bug

* **Handle existing file errors when downloading files**

  Bolt now handles existing file errors raised when creating the
  destination directory for file downloads. Previously, if a file
  already existed somewhere on the destination directory path, Bolt
  would raise an error with a full backtrace.