(GH-2012) Print objects using out::message
This allows users to pass any data type to the `out::message()` plan
function and send it to the configured outputter for printing. This
adds a formatting function to the outputter that all messages pass
through, which formats known Bolt datatypes and classes nicely for
readable printing. Any unhandled types are stringified and passed to the
outputter. Printing is still left entirely to the outputter, but
formatting is shared among all outputter that inherit from the base
outputter class `Bolt::Outputter`. This ensures a consistent experience
across outputters.

Closes #2012

!feature

* **Print objects using out::message()** ([#2012](https://github.com/puppetlabs/bolt/issues/2012))

  Users can now print any valid data type using the plan function
  `out::message()`.