(GH-2004) Add 'bolt plan new' CLI command
This adds a new `bolt plan new` command for generating a project-level
YAML plan. The command accepts a single argument, the name of the plan
that should be generated. If the given name is a valid plan name, and
the plan does not yet exist, Bolt will generate the necessary
directories under `<project>/plans` as well as the YAML plan file. Bolt
will then display the path to the plan, as well as commands to show and
run the plan.

The generated plan includes a single parameter and a couple simple plan
steps, comments describing each portion of the plan, and a link to the
'Writing YAML plans' documentation. The intent behind the generated plan
is to show users the structure of a YAML plan and provide helpful
information to get started with writing a plan.

This also fixes a bug where multi-line strings could be accepted as
project names, due to incorrect anchors in the regex (`^/$` instead of
`\A/\z`).

!feature

* **Create new project-level YAML plans with `bolt plan new`**
  ([#2004](https://github.com/puppetlabs/bolt/issues/2004))

  Users can now quickly get started with writing a new project-level
  YAML plan using the `bolt plan new` command. The command accepts a
  single argument, the name of the plan to be generated, and creates the
  necessary directories and file in the project's `plans` directory.

  > **Note:** This feature is experimental and is subject to changes.

!bug

* **Reject multi-line project names**

  Previously, Bolt would accept a multi-line string as a project name,
  causing multiple errors. Bolt will now reject multi-line strings as
  project names.