(maint) Report screen_view for `bolt plan convert` command
Previously, the code for this command was executing and exiting *before*
the point where we would send a data point about its invocation to
google analytics. That meant it appeared that nobody had ever used this
command. We now fold it into an existing case expression after the
analytics report has been sent.

!no-release-note