(GH-1850) Make examples section in docs
This adds a new 'examples' section to the docs, below 'Using Bolt with
Puppet in the site tree. It moves the 'automating windows targets' and
'tig' example docs to be in this section, and provides a brief
description of what each example exemplifies on the main examples page.
This provides a central place for users to look for examples, and
clarifies the site tree.

Closes #1850

!no-release-note