(bugfix) Don't remove Bolt modulepaths from the modulepath when listing
When we print the modulepath for users when listing tasks or plans, we
previously would calculate the modulpath by subtracting the
Bolt-packaged modulepaths `boltlib` and `modules` from the full
modulepath. This erroneously removes `modules` from the modulepath if it
is listed in the user's path though, which is confusing. This modifies
the modulepath to be calculated from the original modulepath the user
supplied, so that any supplied modulepaths aren't incorrectly removed.

!bugfix

* **Include modules/ in displayed modulepath if it's in the user-configured modulepath**

  Bolt now includes `<bolt-installation-directory>/modules` in the displayed
  modulepath if the user has the path as a component of their configured
  modulepath. If installed as a package on *nix,
  bolt-installation-directory would be `/opt/puppetlabs/bolt/lib/ruby/gems/2.5.0/gems/bolt-x.y.z/`.