(bug) Return empty hash when analytics file is empty
When Bolt loads the anayltics config file, it uses the `YAML.load`
method. If the file is empty, or otherwise cannot be parsed as YAML,
this method returns `false`. This causes an error when Bolt checks the
loaded config for the `disabled` key, as the returned value is not a
hash.

This updates the `Bolt::Analytics.load_config` method to return an empty
hash when the analytics config file cannot be parsed as YAML.

!bug

* **Do not error when analytics configuration file is empty**

  Empty analytics configuration files would cause Bolt to raise a
  `NoMethodError`. Now, if an analytics configuration file is empty,
  Bolt will instead rewrite the file.