(GH-2051) Always set `bolt_project` Puppet setting
This adds a method load_as_module? to the Bolt::Project class to use
in Puppet to determine whether to load the project as a module.
Previously, bolt_project would only be set if the project was named,
which isn't necessary for a valid project, and Puppet would only load
the project as a module if the setting was set. Puppet can now use
load_as_module? to determine whether to load the project as a module,
which lets us always set the bolt_project Puppet setting. This lets us
use that setting to access the project from within Bolt plan functions
for things like getting the project path to set the destination to
download files to.

This PR also removes the now-unnecessary bolt_project_data Puppet
setting, which was an interim means of accessing the project path from
plan functions.

Closes #2051

!no-release-note