(GH-2078) Add Bolt guide pages
This adds a new CLI command and PowerShell cmdlet  which display information
about various Bolt topics. The guides are read from text files that are
saved in the `guides` directory, and are only loaded when a specific
guide is being requested.

To show a list of available guides:

- **Unix shell command**

  ```
  $ bolt guide
  ```

- **PowerShell cmdlet**

  ```
  Get-Help about_bolt_*
  ```

To view the guide for a specific topic:

- **Unix shell command**

  ```
  $ bolt guide inventory
  ```

- **PowerShell cmdlet**

  ```
  Get-Help about_bolt_inventory
  ```

This also adds an analytics event for both known guides and unknown
guides. This allows Bolt to collect data about which guides users are
viewing and which guides users may expect to be available.

!feature

* **View information about Bolt concepts and features from the CLI**
  ([#2078](https://github.com/puppetlabs/bolt/issues/2078))

  Bolt can now display information about various Bolt features and
  concepts with the new CLI command `bolt guide`.