(bug) Pass YAML plan message argument in array
This fixes a bug in the YAML plan message step, which was passing
arguments to the `out::message` plan function incorrectly. Previously,
the message was passed directly to the plan function. If the message was
an object such as an array or hash, the object would be splatted and
passed to the `out::message` plan function as multiple arguments. Now,
the message is passed as a single element array to the plan function,
allowing the function to be called correctly.

!bug

* **Correctly pass objects to YAML plan message step**

  Bolt was incorrectly passing objects used in the YAML plan `message`
  step, resulting in an error. Any objects used in YAML plan `message`
  steps are now correctly passed and printed to the console.