(GH-2081) Update bolt project migrate
This updates the `bolt project migrate` command to rename `bolt.yaml` to
`bolt-project.yaml` in addition to it's previous behavior of updating a
version 1 inventory to version 2. There are two changes that support
this addition:
- The `migrate_project` handler method moved to be in a new class
  `Bolt::ProjectMigrate`. This is purely organizational to avoid
  overloading the `Bolt::CLI` class.
- Backup any files before modifying them. We now backup inventory and config
  files to `<project>/.bolt-bak/<filename>.<date>.bak` before rewriting
  them with updated content. This is to preserve any valuable formatting,
  and to ensure that the file is recoverable should something go wrong
  while writing out the new file.

When migrating a project, any transport configuration in `bolt.yaml` is
removed from that file and merged with the 'all' group transport
configuration in the inventory, with the inventory config taking
precedence. `bolt.yaml` is then renamed to `bolt-project.yaml`. If the
inventory file doesn't already exist and transport config needs to be
added, the inventory file is created. If no config needs to be moved,
the inventory isn't created.

If `bolt-project.yaml` already exists this will not overwrite it. If
`bolt.yaml` does not exist it will not create a `bolt-project.yaml`.

Closes #2081

!feature

* **bolt project migrate now updates project files** ([#2081](https://github.com/puppetlabs/bolt/issues/2081))

  The `bolt project migrate` command will now update bolt.yaml to
  bolt-project.yaml and move transport configuration to inventory.yaml.