(GH-2090 Improve PowerShell parameter template
This commit changes the powershell parameter template to properly handle detecting the last parameter so the correct amount of commas are inserted.

Previously the version parameter was used as a 'last' parameter so the commas lined up, because the erb templating did not keep track of how many parameters were being written. Now the array of parameters is properly joined with a command and the correct amount of newlines.

This commit also changes the template to only includes the `-version` parameter if it is creating a `get` function. The `invoke` functions require input, and so will not allow just the `-version` parameter to be passed.

!no-release-note