(maint) Add debug and verbose to mapped pwsh params
The debug and verbose params are purposely left out of the parameter parsing and templating steps because they come with each PowerShell function, but we need to ensure they are added to the global @mapped_options so that the final PowerShell commandline builder function knows those options exist.

Doing this another way would require us adding filters to the ERB template and makes that more complicated and fragile. Doing it this way does add some duplicated code but keeps the intent clear and in one place.

!no-release-note