(GH-2082) Add 'module install' command to install module dependencies
This adds a new Bolt command `bolt module install` that installs a list of
modules declared in the project's configuration file, as well as all of
their dependencies. This command will resolve all dependencies for the
declared modules, generate a Puppetfile with specifications for each
module, and then install the modules to the project's `.modules`
directory.

As part of this work, the project configuration file,
`bolt-project.yaml`, now accepts a new `modules` key. This key accepts
an array of module declarations, where each declaration is a map that
must include the module's name under the `name` key. The `name` key is
the only supported key at this time.

The installed modules can be found in a new `.modules` directory within
the project. This directory is not currently added to the modulepath,
though users can configure their modulepath to include it if they want.

Depending on the modules declared in the project configuration file, and
whether the project has an existing Puppetfile, this command will behave
differently:

- **No module declarations**

  If the project configuration file `bolt-project.yaml` does not include
  a `modules` key, the command will do nothing.

- **No existing Puppetfile**

  If there are module declarations, but no existing Puppetfile, Bolt
  will resolve the module declarations, generate a new Puppetfile, and
  install the modules.

- **Existing Puppetfile with matching module declarations**

  If there are module declarations, and an existing Puppetfile has
  specifications for each of the module declarations, then Bolt will not
  resolve the module declarations and instead just reinstall the
  Puppetfile.

- **Existing Puppetfile with missing module declarations**

  If there are module declarations, and an existing Puppetfile does not
  have specifications for each of the module declarations, then Bolt
  will raise an error indicating that the Puppetfile may not be managed
  by Bolt.

  If a user wishes to override this behavior, they can run the command
  with the `--force` option, which will force Bolt to resolve the
  modules and overwrite the Puppetfile with a new Puppetfile:

  ```shell
  $ bolt module install --force
  ```

This work introduces a few new classes:

- `Bolt::Puppetfile`
   This manages the logical contents of a Puppetfile and includes
   methods for parsing a Puppetfile, resolving module dependencies,
   and writing a Puppetfile.

- `Bolt::Puppetfile::Installer`
   This installs modules from a Puppetfile.

- `Bolt::Puppetfile::Module`
   This represents a module specification. It includes a comparator
   and some convenience methods.

Lastly, this refactors the `Bolt::CLI#install_puppetfile` and
`Bolt::CLI#initialize_project` methods to use these new classes.

---

**Testing this feature**

This feature is feature flagged. To enable the `bolt module install`
command, set the environment variable `BOLT_MODULE_FEATURE`.

!no-release-note