(PE-29929) Add /tasks endpoint to bolt-server
This commit adds a new endpoint to list all tasks to bolt-server,
similar to the /plans endpoint that already exists. Tasks are
returned as an array where each element in the array is of the
form `{"name":"<TASK_NAME>"}`. This commit also adds a spec test for
the new endpoint.

!no-release-note