(maint) Properly mark non-optional YAML plan parameters in usage info
Previously, YAML plans without a default value were always marked as
optional (surrounded with brackets) in the usage info displayed in
`bolt plan show <plan>`. That happened because we always populate a
default_value field for those parameters, even if it happens to be nil.
Since nil is not a valid value in Puppet, we now explicitly check for it
instead of just using the presence of the key.