(GH-2337) Validate config using schemas
This adds a new `Bolt::Config::Validator` class which validates config
against a schema. The validator will recurse through the config and
schema, validating each value against the corresponding definition in
the schema. If validation fails for a given value, the validator will
add a message to a list of errors. After the entire config has been
validated, and if any invalid values were detected, the validator will
raise an error that lists all of the invalid values.

The validator provides basic validation for config values, such as type
checking, ensuring required keys are present, and that items in an array
are unique. It does not provide more complex validation such as ensuring
a file exists at a given path or that a URL can be reached.

!feature

* **Improved config validation**
  ([#2337](https://github.com/puppetlabs/bolt/issues/2337))

  Bolt now validates config files against Bolt's schemas and indicates
  which config file an error is found in.