(docs) Update run_* function calls in 'Writing plans' docs
This updates several `run_*` function calls in the example plans of the
'Writing plans' doc to use a consistent format for specifying plan,
task, and parameter names. Previously, the doc was inconsistent in how
it showed specifying plan, task, and parameter names, switching between
quoted and unquoted names.

This also fixes a `run_plan` function call that was specifying the
`**puppetdb_fact**` plan, instead of `puppetdb_fact`.

!no-release-note