(GH-2022) Pickup puppet agent 4.1.1
Puppet agent 4.1.1 includes several updates we'd like to pick up,
including:
- Stopping agent runs if Puppet is upgraded mid-run
- Adding a puppet_agent::run plan to run the agent on remote targets

This allows users to run the agent without Bolt erroring due to a
non-zero exit code.

This additionally fixes up a Windows test that upgrades the agent.
The puppet_agent module will now error when attempting to upgrade on
Windows while the service is running, so we now stop the service before
upgrading.

Closes #2022

!feature

* **Add puppet_agent::run plan to run the agent** ([#2022](https://github.com/puppetlabs/bolt/issues/2022)

  The puppet_agent::run plan will run the agent if it's available and
  returns a resultset including agents that failed and the results from
  runs that succeeded.