(GH-2039) Add plan and param descriptions to converted plans
This modifies the plan transpiler to include plan description and plan
parameter descriptions in the converted plan output when converting from
a YAML plan to a Puppet plan. This update is simple as plan and
parameters descriptions are already available from the YamlPlan object,
so just adds them in the proper puppet-strings format to the Puppet
plan. The converted plan warning is part of the plan description by
default.

Closes #2039

!feature

* **Plan conversion maintains plan and parameter descriptions** ([2039](https://github.com/puppetlabs/bolt/issues/2039))

  Converting a YAML plan to a Puppet plan will now preserve plan and
  parameter descriptions, so that `plan show` output is the same for the
  YAML plan as the converted Puppet plan.