(GH-1995) Log where projects are loaded from
This will now log the path that a project is loaded from at `info`
level. This also moves the message listing loaded config files to
`debug`. Lastly this adds a message when Bolt is searching for a project
and doesn't find one at the path stating that the path won't be loaded
as a project.

This also does some restructuring of how logs are gathered from project
and config loading and passed back to the CLI to be logged. Because the
project and config are loaded before the logger is configured (because
log configuration is in the config), we need to store up those logs and
pass them back to the CLI so they can be logged once everything is set
up. Previously we just had a `warnings` instance variable on the config
class that inherited from the project class, got read and logged. The
same flow is true, but now instead of an array of messages to log at
`warn` level there's a new instance variable `logs` which stores an
ordered array of hashes, where each hash is a log level and it's
message. This lets us print the logs in order at the correct log level
in the CLI.

Closes #1995

!no-release-note