(GH-2118) Fix Invoke-BoltTask -params
The -params parameter was originally typed an [object] to allow both string action=status name=spooler and hashtable values @{ action = 'status'; name = 'spooler' }. This succeeded with hashtables but failed with strings because PowerShell parses action=status name=spooler into two groups of parameter-value pairs: action=status and name=spooler. PowerShell then (rightly) fails to bind the second set of params to any of the existing parameter sets and produces an error. The fix for this is to make -params have the ValueFromRemainingArguments attribute which will pull in all of the task params and store it in -params.

This fix however falls into a subtle trap in the switch statement in Get-BoltCommandLine. In PowerShell switch statements unwrap arrays and since we 'fixed' -params to now be an array, it takes action=status name=spooler and makes --params action=status --params name=spooler, which bolt (rightly) fails to process. To fix this an if/else statement is used. While this is more verbose than the switch statement, it ensures correct parsing of arrays, and presumably allows for more parameter types to be handled correctly in the future.

As a result of this investigation, I found/rediscovered the $params variable used in Invoke-BoltCommandLine. Since Invoke-BoltTask uses a parameter called -params, and parameters are variables, there are two 'uses' of the variable $params. This could cause a subtle bug where the value of $params is reassigned before it could be used, but we avoid this because we use $PSBoundParameters instead of individual parameters when building the commandline in Get-BoltCommandLine. To be safe, I renamed $params to be $executionParams to avoid any name clashes.

!bug

* **Correctly handle array parameters**

   Change the switch statement in the Get-BoltCommandline function to an if/else statement to properly handle parameters that are arrays that are not supposed to be unwrapped.