(Docs) Update side nav titles for plans and tasks
This gives the plans and tasks parent pages short titles in the
navigation, while preserving the longer titles on the actual pages.
I've had some feedback that the long titles make it harder for users
to find these pages. This should make it easier to quickly find the
docs in the sidebar while retaining the longer titles on the pages,
which I feel offer important contextual information.

!no-release-note