(GH-2120) Disable log files when set to "disable"
We now allow users to set a log config to the string "disable" to
indicate that the file should be disabled if it was defined at a
previous level of the config hierarchy. This also allows users to
disable the default bolt-debug.log file.

This behavior actually worked already, but only incidentally as we
happened to skip over any logs whose config wasn't a Hash. We now
explicitly check for the string "disable" and otherwise fail if the
value isn't a Hash.

!feature

  * **Disable log files by setting them to `disable`**
     ([#2120](https://github.com/puppetlabs/bolt/issues/2120))

    Log files can now be disabled if they were set at a previous level
    of the hierarchy. This also allows the default `bolt-debug.log` file
    to be disabled.