(maint) Ensure logging is always initialized before logging
When using Bolt as a library, it was possible to create Bolt objects
that had loggers which weren't safe to be used as the levels
corresponding to the methods they call hadn't been defined. In
particular, this would cause a NoMethodError any time Bolt tried to log
something at trace level.

We now have a `Bolt::Logger.logger` method which wraps the
`Logging.logger` method and ensures that we have first called
`initialize_logging` to create the appropriate log levels. This means
consumers of Bolt no longer need to worry about doing that themselves.

!no-release-note