(GH-2116) Handle file writing errors
This cleans up how we handle failures when Bolt fails to write a file to
the current project. We now attempt to create the project directory if
it is the default directory and does not exist. If that creation fails,
we warn and continue without a Bolt project - this effectively means
that Bolt has default configuration, and that logfiles and rerun files
will not be written. If any individual file can't be written to the
project, we warn once and continue. If multiple files can't be written,
we warn for the first file that isn't writeable and not for subsequent
files.

When users explicitly configure a logfile that is unwriteable we now
error.

When users explicitly configure a project that is unwriteable, we now
error.

Closes #2116

!bug

* **Handle project file writing errors more gracefully** ([#2116](https://github.com/puppetlabs/bolt/issues/2116))

  Bolt will now warn and continue executing when it fails to write files
  to the active project. Additionally, any user-specified file data that
  fails to be written to will error.