Extract Puppet interaction to a manager service
This pulls out the code from PAL that handles interacting with Puppet
and building a compiler into a new CompilerService class. That class
builds a compiler context in a dedicated thread and accepts operations
to perform via a work queue, posting results via a promise.

This allows us to only instantiate a single compiler to perform multiple
operations, rather than getting a new one each time we need to use
Puppet.

It also helps separate the code that knows how to interact with Puppet
on a basic level (CompilerService) from the code that knows how to do the particular
operations we care about in Bolt (Bolt::PAL itself).