(GH-2108) Don't report project as conflicting with itself
The previous algorithm for conflict detection checked the modules list
for a module with the same name as the project. This always reported a
conflict, since we include the project as a module and it's always
present. A sensible alternative would be to check for a module with the
same name as the project but a different path. However, *because* of the
very shadowing we're trying to detect, the "real" module doesn't appear
in the modules list in the first place.

We now scan the modulepath explicitly, which excludes the project,
rather than using the final result after Puppet has already removed the
shadowed module.

We also bypass the check entirely if we're not loading the project as a
module. This is just a short-circuit that doesn't change the correctness
of the method, since no module would ever have matched the name `nil` to
issue this warning in that case.

!no-release-note